agent_folder/requirements.txt
python-dotenv==1.0.1
rich==13.7.1
pydantic==2.7.4
langchain==0.2.16
langchain-community==0.2.12
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
email-validator>=2.0
transformers==4.42.4
torch>=2.1.0
faiss-cpu==1.8.0.post1
requests>=2.31.0
sentence-transformers==3.0.1
openai==1.43.0
httpx==0.27.0
fastapi==0.111.0
uvicorn==0.30.0